# NEON optimizations for range/range2