It is very likely that these optimizations apply to the x86 side.

 - Split error counters, join at the end
   - Prevents `vqtbl1q`'s latency from stalling the loop.
 - Remove redundant `vqsubq`s
   - Merge `prev_first_len` and `first_len`, then saturate sub that.

Default benchmark, Clang 10.0.1, Google Pixel 2 XL (Snapdragon 835):
 - range: 1069 MB/s -> 1136 MB/s (6% speedup)
 - range2: 1202 MB/s -> 1267 MB/s (5% speedup) 

Also fixes a compilation error when compiled on non-AVX2 targets.