@easyaspi314 2020-09-12
After testing, manually interleaving `range2_neon` gives a true 15% speedup compared to `master` (not the fake one that happened when Clang optimized out error2).