@easyaspi314 2020-09-12
Another thing that has comparable speed (needs more benchmarking) is replacing 

```c
    uint8x16_t range = vqtbl1q_u8(first_range_tbl, high_nibbles);
```
with
```c
    /* range = (input < 0xC0) ? 0 : 8 */
    uint8x16_t range = vclzq_u8(vcltq_u8(input, const_c0));
```