@easyaspi314 2020-09-12
This optimization also applies to x86,  although parallel error checking is detrimental without AVX's extra registers. 

Once this is done, I will open a PR for the x86 side.