@easyaspi314 2020-09-12
A side note: I was messing around and apparently, a quadruple unroll gets 1830 MB/s. Only aarch64 can pull it off because it can hold everything in the registers, and any more than 4 seems to slow down.

I presume this could also be done with a double AVX2 loop.