@cyb70289 2020-09-14
Observed similar improvement on several Arm64 CPUs from different vendors. Thank you @easyaspi314 !